small refactor to keep the code cleaner
