+++ src/main/kotlin/com/github/ramonvermeulen/dbtToolkit/services/ManifestService.kt
@@ -1,6 +1,8 @@
 package com.github.ramonvermeulen.dbtToolkit.services
 
-import com.google.gson.JsonArray
+import com.github.ramonvermeulen.dbtToolkit.models.Edge
+import com.github.ramonvermeulen.dbtToolkit.models.LineageInfo
+import com.github.ramonvermeulen.dbtToolkit.models.Node
 import com.google.gson.JsonObject
 import com.google.gson.JsonParser
 import com.intellij.openapi.components.Service
@@ -10,90 +12,11 @@ import java.io.File
 import java.util.concurrent.locks.ReentrantLock
 import kotlin.concurrent.withLock
 
-data class Node(
-    val id: String,
-    val tests: MutableList<String> = mutableListOf(),
-    val isSelected: Boolean = false,
-    val relativePath: String,
-) {
-    override fun equals(other: Any?): Boolean {
-        if (this === other) return true
-        if (javaClass != other?.javaClass) return false
-
-        other as Node
-
-        // isSelected is not considered in the equality check
-        // to prevent unnecessary re-renders of the UI lineage graph
-        if (id != other.id) return false
-        if (tests != other.tests) return false
-        if (relativePath != other.relativePath) return false
-
-        return true
-    }
-
-    override fun hashCode(): Int {
-        var result = id.hashCode()
-        result = 31 * result + tests.hashCode()
-        result = 31 * result + relativePath.hashCode()
-        return result
-    }
-}
-
-fun Node.toJson(): JsonObject {
-    val json = JsonObject()
-    json.addProperty("id", id)
-    json.addProperty("isSelected", isSelected)
-    json.add("tests", JsonArray().apply { tests.forEach { add(it) } })
-    json.addProperty("relativePath", relativePath)
-    return json
-}
-
-data class Edge(
-    val parent: String,
-    val child: String,
-)
-
-fun Edge.toJson(): JsonObject {
-    val json = JsonObject()
-    json.addProperty("parent", parent)
-    json.addProperty("child", child)
-    return json
-}
-
-data class LineageInfo(
-    val nodes: MutableSet<Node>,
-    val edges: MutableSet<Edge>,
-) {
-    override fun equals(other: Any?): Boolean {
-        if (this === other) return true
-        if (javaClass != other?.javaClass) return false
-
-        other as LineageInfo
-
-        if (nodes != other.nodes) return false
-        if (edges != other.edges) return false
-
-        return true
-    }
-
-    override fun hashCode(): Int {
-        var result = nodes.hashCode()
-        result = 31 * result + edges.hashCode()
-        return result
-    }
-}
-
-fun LineageInfo.toJson(): JsonObject {
-    val json = JsonObject() // t.b.d.
-    json.add("nodes", JsonArray().apply { nodes.forEach { add(it.toJson()) } })
-    json.add("edges", JsonArray().apply { edges.forEach { add(it.toJson()) } })
-    return json
-}
 
 @Service(Service.Level.PROJECT)
 class ManifestService(private var project: Project) {
     private val lineageInfoService = project.service<LineageInfoService>()
-    private var settings = project.service<DbtToolkitSettingsService>()
+    private var settingsService = project.service<DbtToolkitSettingsService>()
     private val dbtCommandExecutorService = project.service<DbtCommandExecutorService>()
     private var manifest: JsonObject? = null
     private val manifestLock = ReentrantLock()
@@ -101,15 +24,16 @@ class ManifestService(private var project: Project) {
     @Synchronized
     private fun parseManifest() {
         dbtCommandExecutorService.executeCommand(listOf("parse", "--no-partial-parse"))
-        val file = File("${settings.state.settingsDbtTargetDir}/manifest.json")
+        val file = File("${settingsService.state.settingsDbtTargetDir}/manifest.json")
         if (file.exists()) {
             manifest = JsonParser.parseString(file.readText()).asJsonObject
         } else {
             println("Manifest file does not exist")
-            println("looked in the following path: ${settings.state.settingsDbtTargetDir}/manifest.json")
+            println("looked in the following path: ${settingsService.state.settingsDbtTargetDir}/manifest.json")
         }
     }
 
+    // todo(ramon) reconsider this implementation, might be improved
     private fun getCompleteLineageForNode(node: String) {
         val visited = mutableSetOf<String>()
         val nodes = mutableSetOf<Node>()
@@ -124,17 +48,16 @@ class ManifestService(private var project: Project) {
             if (node in visited) return
             val tests = mutableListOf<String>()
             visited.add(node)
-            var nodePath = ""
 
             val children = manifest!!.getAsJsonObject("child_map").getAsJsonArray(node)
             val parents = manifest!!.getAsJsonObject("parent_map").getAsJsonArray(node)
 
-            if (manifestNodes.has(node)) {
-                nodePath = manifestNodes.getAsJsonObject(node).get("original_file_path").asString
-            }
-
-            if ("source" in node) {
-                nodePath = manifestSources.getAsJsonObject(node).get("original_file_path").asString
+            val nodePath: String = if (manifestNodes.has(node)) {
+                manifestNodes.getAsJsonObject(node).get("original_file_path").asString
+            } else if ("source" in node) {
+                manifestSources.getAsJsonObject(node).get("original_file_path").asString
+            } else {
+                ""
             }
 
             children?.forEach { child ->
